Fix rebase conflict artifacts and errors
