+++ src/orion/core/cli/__init__.py
@@ -12,7 +12,7 @@
 import logging
 
 from orion.core.cli.base import OrionArgsParser
-import orion.core.utils.module_import
+from orion.core.utils import module_import
 
 log = logging.getLogger(__name__)
 
@@ -20,10 +20,10 @@ log = logging.getLogger(__name__)
 def load_modules_parser(orion_parser):
     """Search through the `cli` folder for any module containing a `get_parser` function"""
     modules = module_import.load_modules_in_path('orion.core.cli',
-                                                 lambda m: hasattr(m, 'get_parser'))
+                                                 lambda m: hasattr(m, 'add_subparser'))
 
-    for m in modules:
-        get_parser = getattr(m, 'get_parser')
+    for module in modules:
+        get_parser = getattr(module, 'add_subparser')
         get_parser(orion_parser.get_subparsers())
