+++ src/orion/core/cli/init_only.py
@@ -21,7 +21,7 @@ def add_subparser(parser):
     """Return the parser that needs to be used for this command"""
     init_only_parser = parser.add_parser('init_only', help='init_only help')
 
-    cli.get_basic_args_group(init_only_parser)
+    cli.get_basic_args_group(init_only_parser, add_branching=True)
 
     cli.get_user_args_group(init_only_parser)
